--- src/K0x.Benchy.DataStorage.Abstractions/IBenchFileLoader.cs
@@ -1,9 +1,0 @@
-﻿using K0x.Benchy.DataStorage.Abstractions.Models;
-
-namespace K0x.Benchy.DataStorage.Abstractions
-{
-    public interface IBenchFileLoader
-    {
-        Task<Bench> LoadAsync();
-    }
-}
