--- src/K0x.Benchy.DataStorage.Abstractions/IBenchProvider.cs
@@ -1,8 +1,0 @@
-﻿using K0x.Benchy.DataStorage.Abstractions.Models;
-
-namespace K0x.Benchy.DataStorage.Abstractions;
-
-public interface IBenchProvider
-{
-    Bench Bench { get; set; }
-}
