+++ src/K0x.Workbench.DataStorage.Abstractions/IBenchFileLoader.cs
@@ -1,0 +1,9 @@
+﻿using K0x.Workbench.DataStorage.Abstractions.Models;
+
+namespace K0x.Workbench.DataStorage.Abstractions
+{
+    public interface IBenchFileLoader
+    {
+        Task<Bench> LoadAsync();
+    }
+}
