+++ src/K0x.Workbench.DataStorage.Abstractions/IBenchProvider.cs
@@ -1,0 +1,8 @@
+﻿using K0x.Workbench.DataStorage.Abstractions.Models;
+
+namespace K0x.Workbench.DataStorage.Abstractions;
+
+public interface IBenchProvider
+{
+    Bench Bench { get; set; }
+}
