+++ src/WpfBlazor/Components/BenchComponent.razor
@@ -1,4 +1,4 @@
-@using K0x.Benchy.DataStorage.Abstractions.Models
+@using K0x.Workbench.DataStorage.Abstractions.Models
 @namespace WpfBlazor.Components
 
 <h3>Bench: @Bench.Label</h3>
