+++ src/WpfBlazor/Components/ToolComponent.razor
@@ -1,4 +1,4 @@
-@using K0x.Benchy.DataStorage.Abstractions.Models
+@using K0x.Workbench.DataStorage.Abstractions.Models
 @using Microsoft.Extensions.Logging
 @using System.Windows
 @namespace WpfBlazor.Components
@@ -40,6 +40,6 @@
                 button: MessageBoxButton.OK,
                 icon: MessageBoxImage.Error);
 
-        } 
+        }
     }
 }
