+++ src/WpfBlazor/Components/TrayComponent.razor
@@ -1,4 +1,4 @@
-@using K0x.Benchy.DataStorage.Abstractions.Models
+@using K0x.Workbench.DataStorage.Abstractions.Models
 @namespace WpfBlazor.Components
 
 <h4>Tray: @Tray.Label</h4>
