+++ src/WpfBlazor/PocPage.razor
@@ -1,7 +1,7 @@
 @page "/poc"
 @using System.Windows
-@using K0x.Benchy.DataStorage.Abstractions
-@using K0x.Benchy.DataStorage.Abstractions.Models
+@using K0x.Workbench.DataStorage.Abstractions
+@using K0x.Workbench.DataStorage.Abstractions.Models
 @using Microsoft.Extensions.Configuration
 @using Microsoft.Extensions.Logging
 @using WpfBlazor.Components
