+++ src/WpfBlazor/Program.cs
@@ -1,6 +1,6 @@
-﻿using K0x.Benchy.DataStorage.Abstractions;
-using K0x.Benchy.DataStorage.Abstractions.Models;
-using K0x.Benchy.DataStorage.JsonFiles;
+﻿using K0x.Workbench.DataStorage.Abstractions;
+using K0x.Workbench.DataStorage.Abstractions.Models;
+using K0x.Workbench.DataStorage.JsonFiles;
 using K0x.DataStorage.JsonFiles;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
